24.01.2026

+++ ProjectManager.Domain/Entities/Assumption.cs
@@ -5,11 +5,11 @@ public class Assumption
     public int Id { get; set; }
     public int SettlementId { get; set; }
     public Settlement Settlement { get; set; }
-    public float MarginGen { get; set; }
-    public float MarginInstall { get; set; }
-    public float Discount { get; set; }
+    public decimal MarginGen { get; set; }
+    public decimal MarginInstall { get; set; }
+    public decimal Discount { get; set; }
     public int Maturity { get; set; }
-    public float CompanyCost { get; set; }
-    public float CompanyGuarantee { get; set; }
-    public float Insurance { get; set; }
+    public decimal CompanyCost { get; set; }
+    public decimal CompanyGuarantee { get; set; }
+    public decimal Insurance { get; set; }
 }
